requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
groq==0.4.2
python-dotenv==1.0.1
deepface==0.0.79
python-multipart==0.0.6
opencv-python==4.9.0.80
numpy==1.26.4 